+++ NEWS.md
@@ -1,5 +1,6 @@
 ## version 3.5.2
 ### Features
+- `funnel()` plot to visualize residuals vs the expected sampling distribution for `RoBMA()` and `RoBMA.reg()` models when using the `algorithm = "ss"`
 - `residuals()` method for `RoBMA()` and `RoBMA.reg()` models when using the `algorithm = "ss"`
 
 ## version 3.5.1